Updated ItemController Actions to consider user

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,31 @@
 ﻿using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            var currentUser = _userManager.GetUserAsync(HttpContext.User).Result;
+            var model = _context.Items
+                .Where(u => u.User.Id == currentUser.Id)
+                .ToList();
 
             return View("Index", model);
         }
@@ -27,8 +37,10 @@ namespace WishList.Controllers
         }
 
         [HttpPost]
-        public IActionResult Create(Models.Item item)
+        public IActionResult Create(Item item)
         {
+            var currentUser = _userManager.GetUserAsync(HttpContext.User).Result;
+            item.User = currentUser;
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -36,9 +48,18 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
-            var item = _context.Items.FirstOrDefault(e => e.Id == id);
-            _context.Items.Remove(item);
-            _context.SaveChanges();
+            var currentUser = _userManager.GetUserAsync(HttpContext.User).Result;            
+            var item = _context.Items
+                .FirstOrDefault(e => e.Id == id );
+            if (item.User != currentUser)
+            {
+                return Unauthorized();
+            }
+            if (item != null)
+            {
+                _context.Items.Remove(item);
+                _context.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
     }
